requirements.txt
streamlit
requests
geopandas
fiona
shapely
pandas
openpyxl
xlrd
matplotlib
contextily
python-docx
